Revert "SALTO-6992: Skip calculating workspace-service changes if there are no pending changes (#6942)"
This reverts commit 99be0d04a25ce51c612ed79c6b829268b2f19ab4.
